feat: refactor resources for cleanliness and add support for experiment 53
