+++ src/dcd_mapping/align.py
@@ -6,7 +6,9 @@ import tempfile
 import uuid
 from pathlib import Path
 from typing import Any, Generator, List, Optional
+from urllib.parse import urlparse
 
+import requests
 from Bio.SearchIO import HSP
 from Bio.SearchIO import read as read_blat
 from Bio.SearchIO._model import Hit, QueryResult
@@ -14,10 +16,13 @@ from cool_seq_tool.schemas import Strand
 from gene.database.database import click
 
 from dcd_mapping.lookup import get_chromosome_identifier, get_gene_location
-from dcd_mapping.resources import (
+from dcd_mapping.mavedb_data import (
     LOCAL_STORE_PATH,
+)
+from dcd_mapping.resource_utils import (
+    ResourceAcquisitionError,
     get_mapping_tmp_dir,
-    get_ref_genome_file,
+    http_download,
 )
 from dcd_mapping.schemas import (
     AlignmentResult,
@@ -62,6 +67,33 @@ def _build_query_file(scoreset_metadata: ScoresetMetadata) -> Generator[Path, An
     _write_query_file(query_file, lines)
     yield query_file
     query_file.unlink()
+
+
+def get_ref_genome_file(
+    silent: bool = True, dcd_mapping_dir: Optional[Path] = None
+) -> Path:
+    """Acquire reference genome file in 2bit format from UCSC.
+
+    :param build: genome build to acquire
+    :param silent: if True, suppress console output
+    :param dcd_mapping_dir: optionally declare genome file storage location
+    :return: path to acquired file
+    :raise ResourceAcquisitionError: if unable to acquire file.
+    """
+    url = "https://hgdownload.cse.ucsc.edu/goldenpath/hg38/bigZips/hg38.2bit"
+    parsed_url = urlparse(url)
+    if not dcd_mapping_dir:
+        dcd_mapping_dir = LOCAL_STORE_PATH
+    genome_file = dcd_mapping_dir / Path(parsed_url.path).name
+    # this file shouldn't change, so no need to think about more advanced caching
+    if not genome_file.exists():
+        try:
+            http_download(url, genome_file, silent)
+        except requests.HTTPError as e:
+            msg = f"HTTPError when fetching reference genome file from {url}"
+            _logger.error(msg)
+            raise ResourceAcquisitionError(msg) from e
+    return genome_file
 
 
 def _run_blat(
